+++ autopush/endpoint.py
@@ -726,7 +726,7 @@ class RegistrationHandler(AutoendpointHandler):
                                         message="Invalid authentication")
 
         self.uaid = uaid
-        self.chid = str(uuid.uuid4())
+        self.chid = uuid.uuid4().hex
         d = deferToThread(self.ap_settings.router.get_uaid, uaid)
         d.addCallback(self._return_router_data)
         d.addErrback(self._overload_err)
@@ -744,8 +744,9 @@ class RegistrationHandler(AutoendpointHandler):
         self.start_time = time.time()
         self.add_header("Content-Type", "application/json")
         params = self._load_params()
+        # If the client didn't provide a CHID, make one up.
         if "channelID" not in params:
-            return self._error(400, "Invalid arguments")
+            params["channelID"] = uuid.uuid4().hex
 
         # If there's a UAID, ensure its valid, otherwise we ensure the hash
         # matches up
@@ -755,7 +756,7 @@ class RegistrationHandler(AutoendpointHandler):
                 return self._error(401, "Invalid Authentication")
         else:
             # No UAID supplied, make our own
-            uaid = str(uuid.uuid4())
+            uaid = uuid.uuid4().hex
             new_uaid = True
         self.uaid = uaid
         router_type = params.get("type")
@@ -769,7 +770,7 @@ class RegistrationHandler(AutoendpointHandler):
             d.addCallback(router.register, params.get("data", {}))
             d.addCallback(self._save_router_data, router_type)
             d.addCallback(self._make_endpoint)
-            d.addCallback(self._return_endpoint, new_uaid)
+            d.addCallback(self._return_endpoint, new_uaid, router)
             d.addErrback(self._router_fail_err)
             d.addErrback(self._response_err)
             d.callback(uaid)
@@ -836,7 +837,7 @@ class RegistrationHandler(AutoendpointHandler):
         return deferToThread(self.ap_settings.make_endpoint,
                              self.uaid, self.chid)
 
-    def _return_endpoint(self, endpoint, new_uaid):
+    def _return_endpoint(self, endpoint, new_uaid, router=None):
         """Called after the endpoint was made and should be returned to the
         requestor"""
         if new_uaid:
@@ -847,6 +848,9 @@ class RegistrationHandler(AutoendpointHandler):
                 channelID=self.chid,
                 endpoint=endpoint,
             )
+            # Apply any router specific fixes to the outbound response.
+            if router is not None:
+                msg = router.amend_msg(msg)
         else:
             msg = dict(channelID=self.chid, endpoint=endpoint)
         self.write(json.dumps(msg))