+++ CHANGELOG.rst
@@ -2,7 +2,26 @@
 Autopush Changelog
 ==================
 
-1.7.0 (**dev**)
+1.8.0 (**dev**)
+===============
+
+Features
+--------
+* Server provided SenderID values for GCM router using clients
+  The GCM router will randomly select one of a list of SenderIDs stored in
+  S3 under the "org.mozilla.services.autopush"/"senderids" key. The values can
+  be loaded into S3 either via the S3 console, or by runnig an instance of
+  autopush and passing the values as the "senderid_list" argument. Issue #185
+* REST Registration will now return a valid ChannelID if one is not specified.
+  Issue #182
+
+Bug Fixes
+---------
+
+WebPush
+-------
+
+1.7.0 (2015-10-21)
 ===============
 
 Features